+++ CHANGELOG.md
@@ -47,6 +47,7 @@ Improvements
 - Added 'advanced_parameters' to Plugin Parameters to allow hiding of Parameters
   (in the GUI) which are usually not required.
 - Added an .active_plugin_header property to the WorkflowTree
+- The GenericTree.order_node_ids now also sets the active node again.
 
 Bugfixes
 --------